testing/web-platform/tests/atomics/dedicated-worker-agent/wake/wake-all-on-loc.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/workerhelper.js></script>
<script>

// Copyright (C) 2017 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wake wakes all waiters on a location, but does not
  wake waiters on other locations.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    var check_done = WorkerHelper.check_done(t, 4);
    var worker1 = `
      onmessage = function(e) {
        var ia = e.data[0];
        var rs = Atomics.wait(ia, 0, 0);
        postMessage(rs);
      }
    `;
    var worker2 = `
      onmessage = function(e) {
        var ia = e.data[0];
        var rs = Atomics.wait(ia, 1, 0, 1000);
        postMessage(rs);
      }
    `;
    var result_ok = function(e) {
      var rs = e.data;
      assert_equals(rs, "ok");
      check_done();
    }
    var result_timeout = function(e) {
      var rs = e.data;
      assert_equals(rs, "timed-out");
      check_done();
    }

    var ia = new Int32Array(new SharedArrayBuffer(2*Int32Array.BYTES_PER_ELEMENT));
    var workers = [];
    for (let i = 0; i < 3; i++) {
      workers.push(WorkerHelper.createFromString(worker1, result_ok));
    }
    workers.push(WorkerHelper.createFromString(worker2, result_timeout));
    setTimeout(() => Atomics.wake(ia, 0), 500);
    WorkerHelper.broadcast(workers, ia);
  }, "atomics-dedicated-worker-wake-all-on-loc");
});
</script>
